code cleanup - moved handleEventEdit function to utils.jsx
